Reorganize trending components
